+++ business/random-signal-analysis/src/main/java/com/real/time/systems/service/RandomSignalService.java
@@ -4,5 +4,5 @@ import com.real.time.systems.model.RandomSignalResponseDto;
 
 public interface RandomSignalService {
 
-    RandomSignalResponseDto calculate(int amountOfHarmonic, long pointsAmount, double frequency);
+    RandomSignalResponseDto calculate(int amountOfHarmonic, int pointsAmount, double frequency);
 }
